refactor(wave_dependency_graph): migrate to platform adapter
Closes #303
